backend/requirements.txt
absl-py==2.2.2
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
auto_gptq==0.7.1
bcrypt==4.3.0
beanie==1.29.0
beautifulsoup4==4.13.4
bidict==0.23.1
bitsandbytes==0.45.3
blinker==1.9.0
blis==1.2.0
boltons==25.0.0
CacheControl==0.14.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
comm==0.2.2
comtypes==1.4.10
confection==0.1.5
configparser==7.2.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
datasets==3.4.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
dominate==2.9.1
ecdsa==0.19.1
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
face==24.0.0
Faker==37.1.0
fastapi==0.115.0
ffmpeg==1.4
filelock==3.18.0
fire==0.7.0
firebase-admin==6.7.0
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
flask-cors==5.0.1
Flask-Login==0.6.2
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.6.0
gekko==1.2.1
geocoder==1.38.1
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
gunicorn==22.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
ipykernel==6.29.5
ipython==8.35.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.2
kiwisolver==1.4.8
langchain==0.3.23
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.30
language_data==1.3.0
lazy-model==0.2.0
Levenshtein==0.27.1
libclang==18.1.1
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.4
mistral_inference==1.6.0
ml_dtypes==0.5.1
more-itertools==10.6.0
motor==3.7.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
mss==10.0.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
noisereduce==3.0.3
numba==0.61.0
numpy==1.26.4
ollama==0.4.8
openai==1.69.0
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.16
packaging==24.2
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
peft==0.14.0
pillow==11.0.0
platformdirs==4.3.7
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pvcheetah==2.1.3
pvorca==1.1.1
pvporcupine==3.0.5
py==1.11.0
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20240210
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyjokes==0.8.3
PyJWT==2.10.1
pymongo==4.12.1
PyMsgBox==1.0.9
pynput==1.8.1
pyparsing==3.2.1
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-jose==3.4.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
python-socketio==5.12.1
python-vlc==3.0.21203
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.1
pywhatkit==5.4
pywin32==309
pywinauto==0.6.9
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
ratelim==0.1.6
redis==5.0.8
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rouge==1.0.1
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.15.2
screen_brightness_control==0.24.2
seaborn==0.13.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-parsing==0.1.7
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.2
SQLAlchemy==2.0.31
srsly==2.5.1
srt==3.5.3
stack-data==0.6.3
starlette==0.38.6
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
tf_keras==2.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.6
visitor==0.1.3
vosk==0.3.45
wasabi==1.1.3
watchfiles==1.0.5
Wave==0.0.2
wcwidth==0.2.13
weasel==0.4.1
webrtcvad==2.0.10
websockets==15.0.1
Werkzeug==2.3.6
wikipedia==1.4.0
WMI==1.5.1
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
